+++ src/components/home/announcements/Announcement.tsx
@@ -1,6 +1,6 @@
 import * as React from "react";
 import AnnouncementCard from "./AnnouncementCard";
-import type { tweetData } from "../../../screens/Home";
+import type { tweetData } from "../../../screens/home/Home";
 
 function Announcement(props: { tweets: tweetData[]; vertical: boolean}) {
   return (
