+++ src/screens/resources/SpecificResource.tsx
@@ -1,7 +1,7 @@
 import { useEffect, useState } from "react";
 import { useLocation, useNavigate } from 'react-router-dom';
 import { AngleLeftIcon } from "@patternfly/react-icons";
-import { APIUrl } from "../Home";
+import { APIUrl } from "../home/Home";
 import LogoBar from "../../components/home/LogoBar";
 import { Button } from "@patternfly/react-core";
 import { bottom } from "@patternfly/react-core/dist/esm/helpers/Popper/thirdparty/popper-core";
